Multithreading now seems to work
Managed to get rid of segmentation errors in the threaded column max
function. These errors seem to have been caused by two factors:

1. Use of R functions handling SEXP constructs inside daughter
threads. Possibly including calls to Rprintf() or similar.

2. Freeing memory allocated with realloc within the daughter thread.

So it seem to work; but given 250 seqeunces there is no benefit and
in fact the function works slower with a larger number of threads.
I don't know why this is, but I will try to work out why at a later
stage.