Memory optimised Smith-Waterman max column function
Added a function that determines column maxes for a comparison of
sequences of length n x m using only n * 2 bytes of memory. The idea
is to use this for scanning genomic sequences for repeat sequences.

Combined with a simple peak detection and a running sum to scan the
matching positiions this should enable scanning the genomes for
reasonably long repeats.