Wrappers for memory optimised Smith-Waterman max column function
added wrappers for the memory optimised SW max column function.
And debugged the errors typical of after-dinner coding with a
glass of red wine on a small laptop...

This function makes it possible to scan a genome for matches to a
specific sequence using Smith-Waterman (though without affine gaps).
Since the function does not keep the whole score table, it requires
memory primarily for storing the column scores. That is 4 bytes per
base, which for a human genome would come to 12GB. These days that
is not so much, and it can be radically reduced by performing an
inline peak finding algorithm, in which case it would only need to
keep the peaks.. Not difficult to implement. Maybe for tomorrow night.